+++ lib/trailblazer/activity/dsl/linear/feature/variable_mapping.rb
@@ -25,7 +25,6 @@ module Trailblazer
           def self.steps_for_normalizer
             {
               # In(), Out(), {:input}, Inject() feature
-              "activity.convert_symbol_options"           => Linear::Normalizer.Task(VariableMapping::Normalizer.method(:convert_symbol_options)),
               "activity.normalize_input_output_filters"   => Linear::Normalizer.Task(VariableMapping::Normalizer.method(:normalize_input_output_filters)),
               "activity.input_output_dsl"                 => Linear::Normalizer.Task(VariableMapping::Normalizer.method(:input_output_dsl)),
             }
@@ -40,52 +39,18 @@ module Trailblazer
 
           # Steps that are added to the DSL normalizer.
           module Normalizer
-            # TODO: remove me once {:input} API is removed.
-            # Convert {:input}, {:output} and {:inject} to In() and friends.
-            def self.convert_symbol_options(ctx, non_symbol_options:, output_with_outer_ctx: nil, **)
-              input, output, inject = ctx.delete(:input), ctx.delete(:output), ctx.delete(:inject)
-              return unless input || output || inject
-
-              dsl_options = {}
-
-              # TODO: warn, deprecate etc
-              dsl_options.merge!(VariableMapping::DSL.In() => input) if input
-
-              if output
-                options = {}
-                options = options.merge(with_outer_ctx: output_with_outer_ctx) unless output_with_outer_ctx.nil?
-
-                dsl_options.merge!(VariableMapping::DSL.Out(**options) => output)
-              end
-
-              if inject
-                inject.collect do |filter|
-                  filter = filter.is_a?(Symbol) ? [filter] : filter
-
-                  dsl_options.merge!(VariableMapping::DSL.Inject() => filter)
-                end
-              end
-
-              ctx.merge!(
-                non_symbol_options:           non_symbol_options.merge(dsl_options),
-                input_output_inject_options:  [{input: input, output: output, inject: inject}, dsl_options], # yes, there were {:input} options.
-              )
-            end
-
             # Process {In() => [:model], Inject() => [:current_user], Out() => [:model]}
-            def self.normalize_input_output_filters(ctx, non_symbol_options:, input_output_inject_options: [], **)
+            def self.normalize_input_output_filters(ctx, non_symbol_options:, **)
               in_exts     = non_symbol_options.find_all { |k, v| k.is_a?(VariableMapping::DSL::In) || k.is_a?(VariableMapping::DSL::Inject) }
               output_exts = non_symbol_options.find_all { |k, v| k.is_a?(VariableMapping::DSL::Out) }
               return unless in_exts.any? || output_exts.any?
-
-              deprecate_input_output_inject_option(input_output_inject_options, in_exts, output_exts)
 
               ctx[:in_filters]  = in_exts
               ctx[:out_filters] = output_exts
             end
 
             def self.input_output_dsl(ctx, non_symbol_options:, **options)
-              # no :input/:output/:inject/Input()/Output()/:initial_input_pipeline passed.
+              # no Input()/Output()/:initial_input_pipeline passed.
               return unless ctx[:in_filters] || ctx[:out_filters] || ctx[:initial_input_pipeline]
 
               extension = Linear.VariableMapping(**options) # {in_filters:}, {out_filters:} and {:initial_input_pipeline}.
@@ -100,21 +65,6 @@ module Trailblazer
               ctx.merge!(
                 non_symbol_options: non_symbol_options
               )
-            end
-
-
-            # TODO: remove for TRB 2.2.
-            # FIXME: 2.2
-            def self.deprecate_input_output_inject_option(input_output_inject_options, *composable_options)
-              return unless input_output_inject_options.any?
-              options, _dsl_options = input_output_inject_options
-
-              deprecated_options_count = options.find_all { |(name, option)| option }.count + (options[:inject] ? options[:inject].count - 1 : 0)
-              composable_options_count = composable_options.collect { |options| options.size }.sum
-
-              return if composable_options_count == deprecated_options_count
-
-              Activity::Deprecate.warn Linear::Deprecate.dsl_caller_location, %(You are mixing #{options.inspect} with In(), Out() and Inject().\n#{VariableMapping.deprecation_link})
             end
           end